lots of ideas on getting android working, including much refactoring

+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/EyesAppiumDriver.java
@@ -58,7 +58,7 @@ public class EyesAppiumDriver extends EyesWebDriver {
     }
 
     public double getDevicePixelRatio () {
-        return ((Long) getCachedSessionDetails().get("pixelRatio")).doubleValue();
+        return ((Double) getCachedSessionDetails().get("pixelRatio")).doubleValue();
     }
 
     /**
